Add EE module
feat: high availability and scalability (#290)
feat: user/group dashboard (#377)
feat: cross project code search (#1509)
feat: web terminal to access running jobs (#1511)
feat: able to use separate storage for Git LFS and build artifacts (#1510)

+++ server-core/src/main/java/io/onedev/server/util/DirectoryVersionUtils.java
@@ -25,17 +25,17 @@ public class DirectoryVersionUtils {
 		}
 	}
 	
-	public static int readVersion(File directory) {
+	public static long readVersion(File directory) {
 		return callWithVersionLock(directory.toPath(), () -> {
 			var versionFile = new File(directory, FILE_VERSION);
 			if (versionFile.exists())
-				return Integer.parseInt(FileUtils.readFileToString(versionFile, UTF_8).trim());
+				return Long.parseLong(FileUtils.readFileToString(versionFile, UTF_8).trim());
 			else
-				return 0;
+				return 0L;
 		});
 	}
 
-	public static void writeVersion(File directory, int version) {
+	public static void writeVersion(File directory, long version) {
 		callWithVersionLock(directory.toPath(), () -> {
 			var versionFile = new File(directory, FILE_VERSION);
 			FileUtils.writeStringToFile(versionFile, String.valueOf(version), UTF_8);
